requirements.txt
cffi==1.15.0
cryptography
pycparser==2.21
PyJWT==2.3.0
